// src/app/(app)/teachers/_types/teacher-details.types.ts
export type TeacherDetails = {
  attachments: string[];
  personalData: {
    birthDate: string;
    cpf: string;
    fatherName: string;
    fullName: string;
    maritalStatus: string;
    motherName: string;
    nationality: string;
    rg: string;
    sex: string;
  };
  schoolEmail: string;
};